+++ packages/oapp-evm/contracts/oapp/interfaces/IOAppMapper.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.20;
+
+interface IOAppMapper {
+    function lzMap(bytes calldata _request, bytes calldata _response) external view returns (bytes memory);
+}