+++ packages/ua-devtools-evm-hardhat/src/constants/tasks.ts
@@ -13,3 +13,10 @@ export const SUBTASK_LZ_OAPP_WIRE_CONFIGURE = '::lz:oapp:wire:configure'
 export const SUBTASK_LZ_OAPP_CONFIG_LOAD = '::lz:oapp:config:load'
 
 export const TASK_LZ_OWNABLE_TRANSFER_OWNERSHIP = 'lz:ownable:transfer-ownership'
+
+export const TASK_LZ_READ_RESOLVE_COMMAND = 'lz:read:resolve-command'
+export const TASK_LZ_OAPP_READ_WIRE = 'lz:oapp-read:wire'
+export const TASK_LZ_OAPP_READ_CONFIG_GET = 'lz:oapp-read:config:get'
+export const TASK_LZ_OAPP_READ_CONFIG_INIT = 'lz:oapp-read:config:init'
+export const SUBTASK_LZ_OAPP_READ_WIRE_CONFIGURE = '::lz:oapp-read:wire:configure'
+export const TASK_LZ_OAPP_READ_CONFIG_GET_CHANNEL = 'lz:oapp-read:config:get:channel'