+++ packages/create-lz-oapp/CHANGELOG.md
@@ -1,5 +1,29 @@
 # create-lz-oapp
 
+## 0.4.0
+
+### Minor Changes
+
+- 599b3a1: Add UniswapV3 Quote example
+
+## 0.3.1
+
+### Patch Changes
+
+- b63ed66: Add test for OAppRead creation
+
+## 0.3.0
+
+### Minor Changes
+
+- e2395b5: Add OApp Read Example
+
+## 0.2.2
+
+### Patch Changes
+
+- b773736: Fix ref option for native-oft-adapter example
+
 ## 0.2.1
 
 ### Patch Changes