+++ packages/oapp-evm/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @layerzerolabs/oapp-evm
 
+## 0.2.0
+
+### Minor Changes
+
+- 2fd45ca: Add Deploy and Wire for OApp Read
+- e2395b5: Add OApp Read Example
+
+## 0.1.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
 ## 0.0.4
 
 ### Patch Changes