+++ packages/protocol-devtools/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @layerzerolabs/protocol-devtools
 
+## 1.1.0
+
+### Minor Changes
+
+- 2fd45ca: Add Deploy and Wire for OApp Read
+
+### Patch Changes
+
+- @layerzerolabs/devtools@0.4.0
+
+## 1.0.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [aa37daf]
+  - @layerzerolabs/devtools@0.4.0
+
 ## 0.4.3
 
 ### Patch Changes